RODBC-package {RODBC}	R Documentation

ODBC Database Connectivity

Description

Package RODBC implements ODBC database connectivity.

See the package manual for details of installation and use. (This will show up as a vignette, and can be accessed via RShowDoc("RODBC", package="RODBC").)

Details

Two groups of functions are provided. The mainly internal odbc* commands implement low-level access to the ODBC functions of similar name. The sql* functions operate at a higher level to read, save, copy and manipulate data between data frames and SQL tables. Many connections can be open at once to any combination of DSN/hosts.

Author(s)

Michael Lapsley and Brian Ripley

[Package RODBC version 1.3-23 Index]