odbcClose {RODBC}	R Documentation

ODBC Close Connections

Description

Close connections to ODBC databases.

Usage

odbcClose(channel)

## S3 method for class 'RODBC'
close(con, ...)

odbcCloseAll()

Arguments

channel, con
RODBC connection object as returned by odbcConnect.

...
additional arguments passed from the generic.

Details

odbcClose cleans up and frees resources. It is also the method for the generic function close.

odbcCloseAll closes all open channels (amongst the first 1000 used in the session).

Channels are closed at the end of an R session, and may also be closed by garbage collection if no object refers to them. In both cases a warning is given (but may not be seen at the end of a console session).

Value

Function odbcClose returns a logical indicating if it succeeded, invisibly unless a warning is given.

The close method returns 0 (success) or 1, invisibly.

Author(s)

Michael Lapsley and Brian Ripley

See Also

odbcConnect

[Package RODBC version 1.3-13 Index]